# internal: generify and move ordered.BySliceValues to top-level place