--- a/graph/community/louvain_directed_multiplex_test.go
+++ b/graph/community/louvain_directed_multiplex_test.go
@@ -17,8 +17,8 @@ import (
 	"gonum.org/v1/gonum/floats"
 	"gonum.org/v1/gonum/floats/scalar"
 	"gonum.org/v1/gonum/graph"
-	"gonum.org/v1/gonum/graph/internal/ordered"
 	"gonum.org/v1/gonum/graph/simple"
+	"gonum.org/v1/gonum/internal/order"
 )
 
 var communityDirectedMultiplexQTests = []struct {
@@ -310,7 +310,7 @@ func TestCommunityQDirectedMultiplex(t *testing.T) {
 			got := floats.Sum(q)
 			if !scalar.EqualWithinAbsOrRel(got, structure.want, structure.tol, structure.tol) && !math.IsNaN(structure.want) {
 				for _, c := range communities {
-					ordered.ByID(c)
+					order.ByID(c)
 				}
 				t.Errorf("unexpected Q value for %q %v: got: %v %.3v want: %v",
 					test.name, communities, got, q, structure.want)
@@ -338,7 +338,7 @@ tests:
 					communityOf[n] = i
 					communities[i] = append(communities[i], simple.Node(n))
 				}
-				ordered.ByID(communities[i])
+				order.ByID(communities[i])
 			}
 			resolution := []float64{structure.resolution}
 
@@ -359,7 +359,7 @@ tests:
 
 			// This is done to avoid run-to-run
 			// variation due to map iteration order.
-			ordered.ByID(l.nodes)
+			order.ByID(l.nodes)
 
 			l.shuffle(rnd)
 
@@ -376,7 +376,7 @@ tests:
 						}
 						migrated[i] = append(migrated[i], simple.Node(n))
 					}
-					ordered.ByID(migrated[i])
+					order.ByID(migrated[i])
 				}
 
 				for i, c := range structure.memberships {
@@ -454,7 +454,7 @@ tests:
 				for n := range c {
 					communities[i] = append(communities[i], simple.Node(n))
 				}
-				ordered.ByID(communities[i])
+				order.ByID(communities[i])
 			}
 
 			gQ := QMultiplex(g, communities, weights, []float64{structure.resolution})
@@ -540,7 +540,7 @@ func TestMoveLocalDirectedMultiplex(t *testing.T) {
 				for n := range c {
 					communities[i] = append(communities[i], simple.Node(n))
 				}
-				ordered.ByID(communities[i])
+				order.ByID(communities[i])
 			}
 
 			r := reduceDirectedMultiplex(reduceDirectedMultiplex(g, nil, weights), communities, weights)
@@ -580,9 +580,9 @@ func TestLouvainDirectedMultiplex(t *testing.T) {
 			for n := range c {
 				want[i] = append(want[i], simple.Node(n))
 			}
-			ordered.ByID(want[i])
+			order.ByID(want[i])
 		}
-		ordered.BySliceIDs(want)
+		order.BySliceIDs(want)
 
 		var (
 			got   *ReducedDirectedMultiplex
@@ -616,9 +616,9 @@ func TestLouvainDirectedMultiplex(t *testing.T) {
 
 		gotCommunities := got.Communities()
 		for _, c := range gotCommunities {
-			ordered.ByID(c)
+			order.ByID(c)
 		}
-		ordered.BySliceIDs(gotCommunities)
+		order.BySliceIDs(gotCommunities)
 		if !reflect.DeepEqual(gotCommunities, want) {
 			t.Errorf("unexpected community membership for %s Q=%.4v:\n\tgot: %v\n\twant:%v",
 				test.name, bestQ, gotCommunities, want)
@@ -631,9 +631,9 @@ func TestLouvainDirectedMultiplex(t *testing.T) {
 			if p.parent != nil {
 				communities = p.parent.Communities()
 				for _, c := range communities {
-					ordered.ByID(c)
+					order.ByID(c)
 				}
-				ordered.BySliceIDs(communities)
+				order.BySliceIDs(communities)
 			} else {
 				communities = reduceDirectedMultiplex(g, nil, weights).Communities()
 			}