--- a/graph/encoding/dot/encode.go
+++ b/graph/encoding/dot/encode.go
@@ -14,7 +14,7 @@ import (
 
 	"gonum.org/v1/gonum/graph"
 	"gonum.org/v1/gonum/graph/encoding"
-	"gonum.org/v1/gonum/graph/internal/ordered"
+	"gonum.org/v1/gonum/internal/order"
 )
 
 // Node is a DOT graph node.
@@ -166,7 +166,7 @@ func (p *simpleGraphPrinter) print(g graph.Graph, name string, needsIndent, isSu
 	}
 
 	nodes := graph.NodesOf(g.Nodes())
-	ordered.ByID(nodes)
+	order.ByID(nodes)
 
 	havePrintedNodeHeader := false
 	for _, n := range nodes {
@@ -206,7 +206,7 @@ func (p *simpleGraphPrinter) print(g graph.Graph, name string, needsIndent, isSu
 	for _, n := range nodes {
 		nid := n.ID()
 		to := graph.NodesOf(g.From(nid))
-		ordered.ByID(to)
+		order.ByID(to)
 		for _, t := range to {
 			tid := t.ID()
 			f := edge{inGraph: name, from: nid, to: tid}
@@ -465,7 +465,7 @@ func (p *multiGraphPrinter) print(g graph.Multigraph, name string, needsIndent, 
 	}
 
 	nodes := graph.NodesOf(g.Nodes())
-	ordered.ByID(nodes)
+	order.ByID(nodes)
 
 	havePrintedNodeHeader := false
 	for _, n := range nodes {
@@ -505,13 +505,13 @@ func (p *multiGraphPrinter) print(g graph.Multigraph, name string, needsIndent, 
 	for _, n := range nodes {
 		nid := n.ID()
 		to := graph.NodesOf(g.From(nid))
-		ordered.ByID(to)
+		order.ByID(to)
 
 		for _, t := range to {
 			tid := t.ID()
 
 			lines := graph.LinesOf(g.Lines(nid, tid))
-			ordered.LinesByIDs(lines)
+			order.LinesByIDs(lines)
 
 			for _, l := range lines {
 				lid := l.ID()