--- a/graph/graph_test.go
+++ b/graph/graph_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 
 	"gonum.org/v1/gonum/graph"
-	"gonum.org/v1/gonum/graph/internal/ordered"
 	"gonum.org/v1/gonum/graph/simple"
+	"gonum.org/v1/gonum/internal/order"
 )
 
 type graphBuilder interface {
@@ -263,8 +263,8 @@ func TestCopyWeighted(t *testing.T) {
 func same(a, b graph.Graph) bool {
 	aNodes := graph.NodesOf(a.Nodes())
 	bNodes := graph.NodesOf(b.Nodes())
-	ordered.ByID(aNodes)
-	ordered.ByID(bNodes)
+	order.ByID(aNodes)
+	order.ByID(bNodes)
 	for i, na := range aNodes {
 		nb := bNodes[i]
 		if na != nb {
@@ -277,8 +277,8 @@ func same(a, b graph.Graph) bool {
 		if len(aFromU) != len(bFromU) {
 			return false
 		}
-		ordered.ByID(aFromU)
-		ordered.ByID(bFromU)
+		order.ByID(aFromU)
+		order.ByID(bFromU)
 		for i, va := range aFromU {
 			vb := bFromU[i]
 			if va != vb {