--- a/graph/internal/ordered/doc.go
+++ b/graph/internal/ordered/doc.go
@@ -1,6 +1,0 @@
-// Copyright ©2017 The Gonum Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-// Package ordered provides common sort ordering types.
-package ordered // import "gonum.org/v1/gonum/graph/internal/ordered"