--- a/graph/topo/tarjan.go
+++ b/graph/topo/tarjan.go
@@ -9,8 +9,8 @@ import (
 	"slices"
 
 	"gonum.org/v1/gonum/graph"
-	"gonum.org/v1/gonum/graph/internal/ordered"
 	"gonum.org/v1/gonum/graph/internal/set"
+	"gonum.org/v1/gonum/internal/order"
 )
 
 // Unorderable is an error containing sets of unorderable graph.Nodes.
@@ -30,7 +30,7 @@ func (e Unorderable) Error() string {
 	return fmt.Sprintf("topo: no topological ordering: cyclic components: %v", [][]graph.Node(e))
 }
 
-func lexical(nodes []graph.Node) { ordered.ByID(nodes) }
+func lexical(nodes []graph.Node) { order.ByID(nodes) }
 
 // Sort performs a topological sort of the directed graph g returning the 'from' to 'to'
 // sort order. If a topological ordering is not possible, an Unorderable error is returned