--- a/internal/order/doc.go
+++ b/internal/order/doc.go
@@ -1,0 +1,6 @@
+// Copyright ©2024 The Gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// Package order provides common sorting functions.
+package order // import "gonum.org/v1/gonum/internal/order"