--- a/internal/order/order.go
+++ b/internal/order/order.go
@@ -1,10 +1,12 @@
-// Copyright ©2015 The Gonum Authors. All rights reserved.
+// Copyright ©2024 The Gonum Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package ordered
+package order
 
 import (
+	"cmp"
+	"slices"
 	"sort"
 
 	"gonum.org/v1/gonum/graph"
@@ -15,24 +17,17 @@ func ByID(n []graph.Node) {
 	sort.Slice(n, func(i, j int) bool { return n[i].ID() < n[j].ID() })
 }
 
-// BySliceValues sorts a slice of []int64 lexically by the values of the
-// []int64.
-func BySliceValues(c [][]int64) {
-	sort.Slice(c, func(i, j int) bool {
-		a, b := c[i], c[j]
-		l := len(a)
-		if len(b) < l {
-			l = len(b)
-		}
+// BySliceValues sorts a slice of []cmp.Ordered lexically by the values of
+// the []cmp.Ordered.
+func BySliceValues[S interface{ ~[]E }, E cmp.Ordered](c []S) {
+	slices.SortFunc(c, func(a, b S) int {
+		l := min(len(a), len(b))
 		for k, v := range a[:l] {
-			if v < b[k] {
-				return true
-			}
-			if v > b[k] {
-				return false
+			if n := cmp.Compare(v, b[k]); n != 0 {
+				return n
 			}
 		}
-		return len(a) < len(b)
+		return cmp.Compare(len(a), len(b))
 	})
 }