@jbduncan 2024-03-08
I've just force-pushed a fix to the `internal: generify and move ordered.BySliceValues to top-level place` commit to make it run successfully on AppVeyor. I'll see if I can rebase all other PRs dependent on this one.