@jbduncan 2024-03-11
> This makes the `BySliceValues` function available outside graph. I don't see why we need to do this. Nor do I see why it needs to be made generic. Are there reasons that you think this is needed?

@kortschak Good question! It's because the contents of `BySliceValues` was duplicated in `spatial/vptree/vptree_test.go`, presumably because generics didn't exist when it was first written, and I wanted to avoid having `spatial` depend on `graph`.

So I thought it best to move `BySlicesValues` to an internal package that both `graph` and `spatial` could use, then fix `spatial` in PR https://github.com/gonum/gonum/pull/1955.

What do you think?