@jbduncan 2024-03-11
Anyway, as discussed in https://github.com/gonum/gonum/pull/1956#issuecomment-1988326786, I'm more than happy to double check if new benchmarks are needed and to run them all.