@jbduncan 2024-03-16
> This needs to pick up the changes in master.

Done!

Let me know if my previous comment about why this PR exists makes sense or not. I'd be more than happy to rephrase things and even close this PR if you don't think it's valuable.