@jbduncan 2024-03-23
> This is a subset of #1955. Why are they separate or not a chain?

@kortschak Good question. It completely escaped me that I could merge them into one PR but keep the commits separate. Even though #1957 also depends on this PR, it should just work!

When I get to this PR, I'll copy #1955 over, benchmark everything and let you know when things are ready to review.