@kortschak 2024-04-13
The name proliferation here is making me baulk at this. I am thinking that it is less worse to just bring everything out to internal/ordered.