@jbduncan 2024-04-14
> The name proliferation here is making me baulk at this. I am thinking that it is less worse to just bring everything out to internal/order (note change in name from "ordered" since the user semantics are now different; it was previously "sort orderd by O", while now it would be "order by O" — I should have noticed this earlier).

Okay, that makes a lot of sense!

I've force-pushed some changes which should address this.

I intend to fix the merge conflicts and run the benchmarks when the other PRs are merged in.