@jbduncan 2024-04-14
I've decided to rebase anyway, so I'll run the benchmarks later this evening. :)