@jbduncan 2024-04-19
I've finally found the time to upload my benchmark results.

Strangely, the `AStarUndirected` benchmarks seem to be all over the place, ranging from -13% to +17%. That being said, the absolute times are in microseconds, so I could chalk it up to my machine.

@kortschak What do you think?

```
goos: darwin
goarch: arm64
pkg: gonum.org/v1/gonum/graph/coloring
                                                       │  bench.old   │              bench.new              │
                                                       │    sec/op    │    sec/op     vs base               │
Coloring/empty/Dsatur-8                                  3.504n ±  0%   3.465n ±  0%  -1.13% (p=0.000 n=10)
Coloring/empty/DsaturExact-8                             3.188n ±  0%   3.163n ±  0%  -0.78% (p=0.000 n=10)
Coloring/empty/RecursiveLargestFirst-8                   3.189n ±  0%   3.159n ±  0%  -0.93% (p=0.000 n=10)
Coloring/empty/SanSegundo-8                              3.189n ±  0%   3.166n ±  0%  -0.71% (p=0.000 n=10)
Coloring/empty/WelshPowell-8                             3.188n ±  0%   3.172n ±  0%  -0.49% (p=0.001 n=10)
Coloring/singleton/Dsatur-8                              414.8n ±  0%   409.4n ±  0%  -1.29% (p=0.000 n=10)
Coloring/singleton/DsaturExact-8                         960.5n ±  0%   949.2n ±  0%  -1.18% (p=0.000 n=10)
Coloring/singleton/RecursiveLargestFirst-8               298.8n ±  0%   295.7n ±  0%  -1.04% (p=0.000 n=10)
Coloring/singleton/SanSegundo-8                          415.0n ±  0%   408.8n ±  0%  -1.48% (p=0.001 n=10)
Coloring/singleton/WelshPowell-8                         289.8n ±  0%   286.5n ±  0%  -1.12% (p=0.000 n=10)
Coloring/kite/Dsatur-8                                   3.322µ ±  0%   3.288µ ±  0%  -1.01% (p=0.000 n=10)
Coloring/kite/DsaturExact-8                              18.04µ ±  0%   17.85µ ±  0%  -1.05% (p=0.002 n=10)
Coloring/kite/RecursiveLargestFirst-8                    2.246µ ±  0%   2.232µ ±  0%  -0.65% (p=0.001 n=10)
Coloring/kite/SanSegundo-8                               4.042µ ±  0%   4.007µ ±  0%  -0.87% (p=0.000 n=10)
Coloring/kite/WelshPowell-8                              1.407µ ±  0%   1.396µ ±  0%  -0.78% (p=0.000 n=10)
Coloring/triangle+1/Dsatur-8                             2.101µ ±  0%   2.083µ ±  0%  -0.86% (p=0.000 n=10)
Coloring/triangle+1/DsaturExact-8                        10.17µ ±  0%   10.09µ ±  0%  -0.80% (p=0.001 n=10)
Coloring/triangle+1/RecursiveLargestFirst-8              1.438µ ±  0%   1.428µ ±  0%  -0.70% (p=0.000 n=10)
Coloring/triangle+1/SanSegundo-8                         2.435µ ±  0%   2.417µ ±  0%  -0.78% (p=0.000 n=10)
Coloring/triangle+1/WelshPowell-8                        980.0n ±  0%   972.1n ±  0%  -0.81% (p=0.000 n=10)
Coloring/bipartite_halves/Dsatur-8                       8.015µ ±  0%   7.975µ ±  1%  -0.50% (p=0.022 n=10)
Coloring/bipartite_halves/DsaturExact-8                  33.30µ ±  0%   33.12µ ±  0%  -0.52% (p=0.000 n=10)
Coloring/bipartite_halves/RecursiveLargestFirst-8        3.982µ ±  0%   3.970µ ±  0%  -0.31% (p=0.000 n=10)
Coloring/bipartite_halves/SanSegundo-8                   15.22µ ±  0%   15.15µ ±  0%  -0.48% (p=0.000 n=10)
Coloring/bipartite_halves/WelshPowell-8                  2.394µ ±  0%   2.384µ ±  1%       ~ (p=0.055 n=10)
Coloring/bipartite_alternating/Dsatur-8                  7.885µ ±  0%   7.875µ ±  0%  -0.12% (p=0.041 n=10)
Coloring/bipartite_alternating/DsaturExact-8             33.11µ ±  0%   32.97µ ±  0%  -0.41% (p=0.000 n=10)
Coloring/bipartite_alternating/RecursiveLargestFirst-8   3.950µ ±  0%   3.934µ ±  1%  -0.42% (p=0.027 n=10)
Coloring/bipartite_alternating/SanSegundo-8              15.32µ ±  0%   15.25µ ±  0%  -0.45% (p=0.000 n=10)
Coloring/bipartite_alternating/WelshPowell-8             2.486µ ±  0%   2.482µ ±  1%       ~ (p=0.341 n=10)
Coloring/3/4_bipartite/Dsatur-8                          6.412µ ±  0%   6.401µ ±  0%  -0.17% (p=0.008 n=10)
Coloring/3/4_bipartite/DsaturExact-8                     30.54µ ±  0%   30.41µ ±  0%  -0.42% (p=0.002 n=10)
Coloring/3/4_bipartite/RecursiveLargestFirst-8           3.529µ ±  0%   3.510µ ±  0%  -0.54% (p=0.000 n=10)
Coloring/3/4_bipartite/SanSegundo-8                      13.38µ ±  0%   13.33µ ±  0%  -0.34% (p=0.005 n=10)
Coloring/3/4_bipartite/WelshPowell-8                     2.235µ ±  0%   2.230µ ±  0%  -0.25% (p=0.000 n=10)
Coloring/cubical/Dsatur-8                                8.019µ ±  0%   7.996µ ±  0%  -0.29% (p=0.001 n=10)
Coloring/cubical/DsaturExact-8                           33.51µ ±  0%   33.43µ ±  0%  -0.24% (p=0.009 n=10)
Coloring/cubical/RecursiveLargestFirst-8                 4.056µ ±  0%   4.040µ ±  0%  -0.38% (p=0.000 n=10)
Coloring/cubical/SanSegundo-8                            15.39µ ±  0%   15.33µ ±  1%  -0.36% (p=0.037 n=10)
Coloring/cubical/WelshPowell-8                           2.513µ ±  0%   2.509µ ±  1%       ~ (p=0.515 n=10)
Coloring/HoG33106/Dsatur-8                               19.06µ ±  0%   19.01µ ±  0%  -0.28% (p=0.002 n=10)
Coloring/HoG33106/DsaturExact-8                          73.89µ ±  0%   73.62µ ±  0%  -0.37% (p=0.000 n=10)
Coloring/HoG33106/RecursiveLargestFirst-8                8.845µ ±  0%   8.825µ ±  0%  -0.23% (p=0.001 n=10)
Coloring/HoG33106/SanSegundo-8                           69.14µ ±  0%   68.94µ ±  0%  -0.30% (p=0.015 n=10)
Coloring/HoG33106/WelshPowell-8                          5.000µ ±  0%   4.972µ ±  0%  -0.56% (p=0.000 n=10)
Coloring/HoG41237/Dsatur-8                               217.2m ±  1%   219.2m ±  1%  +0.93% (p=0.001 n=10)
Coloring/HoG41237/DsaturExact-8                          239.9m ±  1%   239.6m ±  1%       ~ (p=0.631 n=10)
Coloring/HoG41237/RecursiveLargestFirst-8                8.899m ±  1%   8.888m ±  1%       ~ (p=0.796 n=10)
Coloring/HoG41237/SanSegundo-8                           506.7m ±  1%   507.4m ±  0%       ~ (p=0.190 n=10)
Coloring/HoG41237/WelshPowell-8                          3.663m ±  1%   3.679m ±  0%       ~ (p=0.165 n=10)
Coloring/exams/Dsatur-8                                  4.660µ ±  0%   4.656µ ±  0%       ~ (p=0.236 n=10)
Coloring/exams/DsaturExact-8                             26.84µ ±  0%   26.80µ ±  0%       ~ (p=0.137 n=10)
Coloring/exams/RecursiveLargestFirst-8                   3.057µ ±  0%   3.052µ ±  0%       ~ (p=0.118 n=10)
Coloring/exams/SanSegundo-8                              7.487µ ±  0%   7.468µ ±  1%  -0.25% (p=0.024 n=10)
Coloring/exams/WelshPowell-8                             1.837µ ±  0%   1.834µ ±  0%       ~ (p=0.107 n=10)
Coloring/exam_scheduler_1/Dsatur-8                       18.01µ ±  0%   17.94µ ±  0%  -0.39% (p=0.001 n=10)
Coloring/exam_scheduler_1/DsaturExact-8                  100.2µ ±  0%   100.0µ ±  0%  -0.18% (p=0.023 n=10)
Coloring/exam_scheduler_1/RecursiveLargestFirst-8        11.41µ ±  0%   11.37µ ±  0%  -0.36% (p=0.000 n=10)
Coloring/exam_scheduler_1/SanSegundo-8                   27.72µ ±  0%   27.72µ ±  0%       ~ (p=0.853 n=10)
Coloring/exam_scheduler_1/WelshPowell-8                  4.740µ ±  0%   4.728µ ±  0%  -0.25% (p=0.005 n=10)
Coloring/exam_scheduler_2/Dsatur-8                       8.370µ ±  0%   8.353µ ±  0%  -0.21% (p=0.001 n=10)
Coloring/exam_scheduler_2/DsaturExact-8                  48.60µ ±  0%   48.39µ ±  0%  -0.43% (p=0.000 n=10)
Coloring/exam_scheduler_2/RecursiveLargestFirst-8        5.192µ ±  0%   5.181µ ±  0%       ~ (p=0.159 n=10)
Coloring/exam_scheduler_2/SanSegundo-8                   12.86µ ±  0%   12.81µ ±  0%  -0.45% (p=0.001 n=10)
Coloring/exam_scheduler_2/WelshPowell-8                  2.754µ ±  0%   2.743µ ±  0%  -0.40% (p=0.005 n=10)
Coloring/Brélaz/Dsatur-8                                 12.12µ ±  0%   12.09µ ±  0%  -0.25% (p=0.004 n=10)
Coloring/Brélaz/DsaturExact-8                            59.96µ ±  0%   59.77µ ±  0%       ~ (p=0.052 n=10)
Coloring/Brélaz/RecursiveLargestFirst-8                  7.611µ ±  0%   7.583µ ±  0%  -0.37% (p=0.009 n=10)
Coloring/Brélaz/SanSegundo-8                             18.52µ ±  0%   18.47µ ±  0%  -0.28% (p=0.005 n=10)
Coloring/Brélaz/WelshPowell-8                            3.721µ ±  0%   3.711µ ±  0%  -0.28% (p=0.014 n=10)
Coloring/Lima_Carmo/Dsatur-8                             7.572µ ±  1%   7.562µ ±  0%       ~ (p=0.078 n=10)
Coloring/Lima_Carmo/DsaturExact-8                        32.98µ ±  0%   33.00µ ±  1%       ~ (p=0.579 n=10)
Coloring/Lima_Carmo/RecursiveLargestFirst-8              4.178µ ±  0%   4.166µ ±  0%  -0.28% (p=0.005 n=10)
Coloring/Lima_Carmo/SanSegundo-8                         11.28µ ±  0%   11.25µ ±  0%  -0.25% (p=0.003 n=10)
Coloring/Lima_Carmo/WelshPowell-8                        2.486µ ±  0%   2.477µ ±  0%  -0.36% (p=0.001 n=10)
Coloring/San_Segundo/Dsatur-8                            12.84µ ±  0%   12.81µ ±  1%       ~ (p=0.280 n=10)
Coloring/San_Segundo/DsaturExact-8                       120.7µ ±  0%   120.2µ ±  0%  -0.39% (p=0.029 n=10)
Coloring/San_Segundo/RecursiveLargestFirst-8             9.222µ ±  0%   9.208µ ±  0%  -0.15% (p=0.045 n=10)
Coloring/San_Segundo/SanSegundo-8                        22.69µ ±  0%   22.67µ ±  0%       ~ (p=0.149 n=10)
Coloring/San_Segundo/WelshPowell-8                       3.621µ ±  1%   3.604µ ±  0%  -0.47% (p=0.011 n=10)
Coloring/tetrahedron/Dsatur-8                            2.391µ ±  0%   2.388µ ±  1%       ~ (p=0.753 n=10)
Coloring/tetrahedron/DsaturExact-8                       13.74µ ±  0%   13.72µ ±  0%  -0.14% (p=0.011 n=10)
Coloring/tetrahedron/RecursiveLargestFirst-8             1.860µ ±  0%   1.854µ ±  0%  -0.32% (p=0.001 n=10)
Coloring/tetrahedron/SanSegundo-8                        3.338µ ±  0%   3.326µ ±  0%  -0.34% (p=0.005 n=10)
Coloring/tetrahedron/WelshPowell-8                       1.069µ ±  0%   1.066µ ±  0%  -0.33% (p=0.000 n=10)
Coloring/triangle/Dsatur-8                               1.487µ ±  0%   1.483µ ±  0%       ~ (p=0.061 n=10)
Coloring/triangle/DsaturExact-8                          7.328µ ±  0%   7.321µ ±  0%       ~ (p=0.117 n=10)
Coloring/triangle/RecursiveLargestFirst-8                1.157µ ±  0%   1.153µ ±  0%  -0.39% (p=0.000 n=10)
Coloring/triangle/SanSegundo-8                           1.828µ ±  0%   1.824µ ±  0%  -0.25% (p=0.026 n=10)
Coloring/triangle/WelshPowell-8                          774.6n ±  0%   771.6n ±  0%  -0.38% (p=0.007 n=10)
Coloring/square/Dsatur-8                                 2.206µ ±  0%   2.202µ ±  0%       ~ (p=0.085 n=10)
Coloring/square/DsaturExact-8                            10.96µ ±  0%   10.93µ ±  0%  -0.22% (p=0.020 n=10)
Coloring/square/RecursiveLargestFirst-8                  1.428µ ±  0%   1.423µ ±  0%  -0.35% (p=0.000 n=10)
Coloring/square/SanSegundo-8                             2.821µ ±  0%   2.814µ ±  0%  -0.27% (p=0.001 n=10)
Coloring/square/WelshPowell-8                            1.024µ ±  1%   1.023µ ±  0%       ~ (p=0.178 n=10)
Coloring/cycle-5/Dsatur-8                                3.238µ ±  1%   3.232µ ±  1%       ~ (p=0.254 n=10)
Coloring/cycle-5/DsaturExact-8                           19.36µ ±  0%   19.31µ ±  0%  -0.26% (p=0.009 n=10)
Coloring/cycle-5/RecursiveLargestFirst-8                 2.122µ ±  0%   2.120µ ±  0%       ~ (p=0.083 n=10)
Coloring/cycle-5/SanSegundo-8                            4.167µ ±  0%   4.155µ ±  0%  -0.29% (p=0.000 n=10)
Coloring/cycle-5/WelshPowell-8                           1.366µ ±  0%   1.365µ ±  0%       ~ (p=0.342 n=10)
Coloring/cycle-6/Dsatur-8                                4.319µ ±  0%   4.314µ ±  0%       ~ (p=0.403 n=10)
Coloring/cycle-6/DsaturExact-8                           18.43µ ±  0%   18.43µ ±  0%       ~ (p=0.868 n=10)
Coloring/cycle-6/RecursiveLargestFirst-8                 2.496µ ±  0%   2.494µ ±  0%       ~ (p=0.381 n=10)
Coloring/cycle-6/SanSegundo-8                            5.607µ ±  0%   5.588µ ±  0%  -0.34% (p=0.004 n=10)
Coloring/cycle-6/WelshPowell-8                           1.673µ ±  0%   1.670µ ±  1%       ~ (p=0.668 n=10)
Coloring/wheel-5/Dsatur-8                                3.458µ ±  0%   3.457µ ±  0%       ~ (p=0.115 n=10)
Coloring/wheel-5/DsaturExact-8                           23.11µ ±  0%   23.06µ ±  0%  -0.21% (p=0.020 n=10)
Coloring/wheel-5/RecursiveLargestFirst-8                 2.393µ ±  0%   2.389µ ±  0%  -0.19% (p=0.011 n=10)
Coloring/wheel-5/SanSegundo-8                            5.205µ ±  0%   5.192µ ±  0%  -0.26% (p=0.015 n=10)
Coloring/wheel-5/WelshPowell-8                           1.402µ ±  0%   1.401µ ±  1%       ~ (p=0.515 n=10)
Coloring/wheel-6/Dsatur-8                                4.867µ ±  0%   4.859µ ±  0%  -0.14% (p=0.024 n=10)
Coloring/wheel-6/DsaturExact-8                           38.28µ ±  0%   38.24µ ±  0%       ~ (p=0.393 n=10)
Coloring/wheel-6/RecursiveLargestFirst-8                 3.323µ ±  0%   3.319µ ±  0%       ~ (p=0.254 n=10)
Coloring/wheel-6/SanSegundo-8                            7.615µ ±  0%   7.602µ ±  0%       ~ (p=0.240 n=10)
Coloring/wheel-6/WelshPowell-8                           1.813µ ±  0%   1.811µ ±  0%       ~ (p=0.693 n=10)
Coloring/wheel-7/Dsatur-8                                6.346µ ±  0%   6.332µ ±  0%  -0.22% (p=0.003 n=10)
Coloring/wheel-7/DsaturExact-8                           39.74µ ±  0%   39.76µ ±  0%       ~ (p=0.529 n=10)
Coloring/wheel-7/RecursiveLargestFirst-8                 3.950µ ±  0%   3.955µ ±  0%       ~ (p=0.183 n=10)
Coloring/wheel-7/SanSegundo-8                            10.21µ ±  0%   10.22µ ±  0%       ~ (p=0.754 n=10)
Coloring/wheel-7/WelshPowell-8                           2.192µ ±  0%   2.186µ ±  0%       ~ (p=0.083 n=10)
Coloring/sudoku_board/Dsatur-8                           21.54m ±  0%   21.46m ±  0%  -0.36% (p=0.000 n=10)
Coloring/sudoku_board/DsaturExact-8                      86.10m ±  0%   86.28m ±  1%       ~ (p=0.218 n=10)
Coloring/sudoku_board/RecursiveLargestFirst-8            3.930m ±  0%   3.936m ±  0%       ~ (p=0.436 n=10)
Coloring/sudoku_board/SanSegundo-8                       25.49m ±  0%   25.51m ±  0%       ~ (p=0.796 n=10)
Coloring/sudoku_board/WelshPowell-8                      691.2µ ±  0%   688.0µ ±  0%  -0.46% (p=0.000 n=10)
Coloring/sudoku_problem/Dsatur-8                         32.66m ±  0%   32.65m ±  0%       ~ (p=0.796 n=10)
Coloring/sudoku_problem/DsaturExact-8                     5.928 ± 50%    5.383 ± 64%       ~ (p=0.481 n=10)
Coloring/sudoku_problem/RecursiveLargestFirst-8          5.886m ±  0%   5.871m ±  0%       ~ (p=0.052 n=10)
Coloring/sudoku_problem/SanSegundo-8                     35.64m ±  0%   35.70m ±  0%       ~ (p=0.353 n=10)
Coloring/sudoku_problem/WelshPowell-8                    963.5µ ±  0%   963.7µ ±  1%       ~ (p=0.315 n=10)
Coloring/cs_shc/Dsatur-8                                 3.365µ ±  0%   3.364µ ±  0%       ~ (p=0.896 n=10)
Coloring/cs_shc/DsaturExact-8                            20.55µ ±  0%   20.59µ ±  0%  +0.18% (p=0.043 n=10)
Coloring/cs_shc/RecursiveLargestFirst-8                  2.392µ ±  1%   2.382µ ±  0%  -0.40% (p=0.001 n=10)
Coloring/cs_shc/SanSegundo-8                             5.141µ ±  0%   5.148µ ±  1%       ~ (p=0.066 n=10)
Coloring/cs_shc/WelshPowell-8                            1.426µ ±  0%   1.425µ ±  0%       ~ (p=0.345 n=10)
Coloring/gis_hc/Dsatur-8                                 3.995µ ±  0%   3.995µ ±  0%       ~ (p=0.752 n=10)
Coloring/gis_hc/DsaturExact-8                            16.91µ ±  0%   16.92µ ±  0%       ~ (p=0.956 n=10)
Coloring/gis_hc/RecursiveLargestFirst-8                  2.318µ ±  0%   2.312µ ±  0%       ~ (p=0.076 n=10)
Coloring/gis_hc/SanSegundo-8                             7.495µ ±  0%   7.493µ ±  1%       ~ (p=0.985 n=10)
Coloring/gis_hc/WelshPowell-8                            1.655µ ±  0%   1.652µ ±  0%       ~ (p=0.272 n=10)
Coloring/gis_shc|rs_shc/Dsatur-8                         2.081µ ±  0%   2.079µ ±  0%       ~ (p=0.134 n=10)
Coloring/gis_shc|rs_shc/DsaturExact-8                    9.811µ ±  0%   9.827µ ±  0%       ~ (p=0.060 n=10)
Coloring/gis_shc|rs_shc/RecursiveLargestFirst-8          1.348µ ±  0%   1.346µ ±  0%       ~ (p=0.052 n=10)
Coloring/gis_shc|rs_shc/SanSegundo-8                     2.713µ ±  0%   2.710µ ±  1%       ~ (p=0.541 n=10)
Coloring/gis_shc|rs_shc/WelshPowell-8                    1.020µ ±  0%   1.018µ ±  0%       ~ (p=0.062 n=10)
Coloring/lf_hc/Dsatur-8                                  6.129µ ±  0%   6.142µ ±  0%       ~ (p=0.109 n=10)
Coloring/lf_hc/DsaturExact-8                             32.56µ ±  0%   32.78µ ±  0%  +0.66% (p=0.003 n=10)
Coloring/lf_hc/RecursiveLargestFirst-8                   3.733µ ±  1%   3.738µ ±  0%       ~ (p=0.343 n=10)
Coloring/lf_hc/SanSegundo-8                              10.33µ ±  0%   10.34µ ±  1%       ~ (p=0.148 n=10)
Coloring/lf_hc/WelshPowell-8                             2.073µ ±  0%   2.072µ ±  0%       ~ (p=0.900 n=10)
Coloring/lf_shc/Dsatur-8                                 4.088µ ±  0%   4.097µ ±  0%  +0.21% (p=0.017 n=10)
Coloring/lf_shc/DsaturExact-8                            16.96µ ±  0%   16.95µ ±  0%       ~ (p=0.382 n=10)
Coloring/lf_shc/RecursiveLargestFirst-8                  2.420µ ±  1%   2.418µ ±  0%       ~ (p=0.146 n=10)
Coloring/lf_shc/SanSegundo-8                             4.460µ ±  0%   4.455µ ±  1%       ~ (p=0.196 n=10)
Coloring/lf_shc/WelshPowell-8                            1.599µ ±  0%   1.597µ ±  0%       ~ (p=0.058 n=10)
Coloring/lfi_hc/Dsatur-8                                 10.50µ ±  0%   10.50µ ±  0%       ~ (p=0.956 n=10)
Coloring/lfi_hc/DsaturExact-8                            67.65µ ±  0%   67.58µ ±  0%       ~ (p=0.271 n=10)
Coloring/lfi_hc/RecursiveLargestFirst-8                  5.940µ ±  0%   5.941µ ±  0%       ~ (p=0.782 n=10)
Coloring/lfi_hc/SanSegundo-8                             14.03µ ±  0%   14.02µ ±  0%       ~ (p=0.055 n=10)
Coloring/lfi_hc/WelshPowell-8                            3.130µ ±  0%   3.126µ ±  0%       ~ (p=0.208 n=10)
Coloring/lfi_shc|slf_shc/Dsatur-8                        6.223µ ±  0%   6.238µ ±  0%       ~ (p=0.072 n=10)
Coloring/lfi_shc|slf_shc/DsaturExact-8                   39.31µ ±  0%   39.22µ ±  0%  -0.24% (p=0.015 n=10)
Coloring/lfi_shc|slf_shc/RecursiveLargestFirst-8         3.552µ ±  0%   3.536µ ±  1%  -0.45% (p=0.018 n=10)
Coloring/lfi_shc|slf_shc/SanSegundo-8                    7.167µ ±  0%   7.158µ ±  0%  -0.13% (p=0.050 n=10)
Coloring/lfi_shc|slf_shc/WelshPowell-8                   2.139µ ±  0%   2.137µ ±  0%       ~ (p=0.145 n=10)
Coloring/no_solo/Dsatur-8                                16.80µ ±  0%   16.78µ ±  0%       ~ (p=0.060 n=10)
Coloring/no_solo/DsaturExact-8                           54.63µ ±  0%   54.64µ ±  1%       ~ (p=0.956 n=10)
Coloring/no_solo/RecursiveLargestFirst-8                 7.693µ ±  0%   7.674µ ±  0%  -0.23% (p=0.012 n=10)
Coloring/no_solo/SanSegundo-8                            25.33µ ±  0%   25.28µ ±  0%       ~ (p=0.127 n=10)
Coloring/no_solo/WelshPowell-8                           4.412µ ±  0%   4.404µ ±  0%  -0.17% (p=0.012 n=10)
Coloring/sl_hc/Dsatur-8                                  8.790µ ±  0%   8.775µ ±  0%       ~ (p=0.093 n=10)
Coloring/sl_hc/DsaturExact-8                             74.83µ ±  0%   74.74µ ±  0%       ~ (p=0.143 n=10)
Coloring/sl_hc/RecursiveLargestFirst-8                   5.267µ ±  0%   5.261µ ±  0%       ~ (p=0.171 n=10)
Coloring/sl_hc/SanSegundo-8                              12.05µ ±  0%   12.03µ ±  0%  -0.20% (p=0.011 n=10)
Coloring/sl_hc/WelshPowell-8                             2.692µ ±  0%   2.690µ ±  0%       ~ (p=0.223 n=10)
Coloring/sl_shc/Dsatur-8                                 4.810µ ±  0%   4.802µ ±  0%       ~ (p=0.100 n=10)
Coloring/sl_shc/DsaturExact-8                            25.46µ ±  0%   25.43µ ±  0%       ~ (p=0.118 n=10)
Coloring/sl_shc/RecursiveLargestFirst-8                  2.996µ ±  0%   2.994µ ±  0%       ~ (p=0.904 n=10)
…